vi-feat: validation schemas and payload keys moved to corresponding folders in packages
