+++ client/src/libs/components/header/header.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import { NavLink } from 'react-router-dom';
 
 import { IconName, IconSize, ButtonType, AppRoute } from 'libs/enums/enums';
-import { DEFAULT_USER_AVATAR } from 'libs/constants/constants';
+import { DEFAULT_USER_AVATAR } from 'packages/user/constants/constants';
 import { userType } from 'libs/prop-types/prop-types';
 import { Button } from '../button/button';
 import { Icon } from '../icon/icon';
