--- client/src/libs/enums/comment/comment.js
@@ -1,1 +1,0 @@
-export { CommentPayloadKey } from 'shared/src/common/enums/enums';
