+++ client/src/libs/enums/enums.js
@@ -1,12 +1,8 @@
 export { ApiPath } from 'shared/src/common/enums/enums';
 export { AppRoute, ENV, StorageKey } from './app/app';
 export { ContentType } from './file/file';
-export { UserPayloadKey } from './user/user';
-export { PostPayloadKey } from './post/post';
-export { ImagePayloadKey } from './image/image';
 export { UseFormMode } from './form/form';
 export { ThreadToolbarKey } from './thread/thread';
-export { CommentPayloadKey } from './comment/comment';
 export { ExceptionMessage } from './exception/exception';
 export {
   ButtonColor,
