--- client/src/libs/enums/image/image.js
@@ -1,1 +1,0 @@
-export { ImagePayloadKey } from 'shared/src/common/enums/enums';
