--- client/src/libs/enums/post/post.js
@@ -1,1 +1,0 @@
-export { PostPayloadKey } from 'shared/src/common/enums/enums';
