+++ client/src/pages/profile/profile.jsx
@@ -1,5 +1,5 @@
 import { useAppForm, useSelector } from 'libs/hooks/hooks';
-import { DEFAULT_USER_AVATAR } from 'libs/constants/constants';
+import { DEFAULT_USER_AVATAR } from 'packages/user/constants/constants';
 import { ImageSize, IconName } from 'libs/enums/enums';
 import { Image } from 'libs/components/image/image';
 import { Input } from 'libs/components/input/input';
