+++ client/src/pages/thread/components/add-comment/add-comment.jsx
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types';
 import { useAppForm, useCallback } from 'libs/hooks/hooks';
-import { ButtonType, CommentPayloadKey } from 'libs/enums/enums';
+import { ButtonType } from 'libs/enums/enums';
+import { CommentPayloadKey } from 'packages/comment/libs/enums/enums';
 import { Button } from 'libs/components/button/button';
 import { Input } from 'libs/components/input/input';
 import { DEFAULT_ADD_COMMENT_PAYLOAD } from './common/constants';
