+++ client/src/pages/thread/components/add-comment/common/constants.js
@@ -1,4 +1,4 @@
-import { CommentPayloadKey } from 'libs/enums/enums';
+import { CommentPayloadKey } from 'packages/comment/libs/enums/enums';
 
 const DEFAULT_ADD_COMMENT_PAYLOAD = {
   [CommentPayloadKey.BODY]: ''
