+++ client/src/pages/thread/components/add-post/common/constants.js
@@ -1,4 +1,4 @@
-import { PostPayloadKey } from 'libs/enums/enums';
+import { PostPayloadKey } from 'packages/post/libs/enums/enums';
 
 const DEFAULT_ADD_POST_PAYLOAD = {
   [PostPayloadKey.BODY]: ''
